# Replace sse2neon with native NEON