--- a/.gitmodules
+++ b/.gitmodules
@@ -1,6 +1,3 @@
 [submodule "cpu_features"]
 	path = cpu_features
 	url = https://github.com/google/cpu_features.git
-[submodule "include/volk/sse2neon"]
-	path = include/volk/sse2neon
-	url = https://github.com/DLTcollab/sse2neon.git