`volk_8u_x4_conv_k7_r2_8u_neonspiral` is the only protokernel that needs sse2neon, which has proven to be a troublesome dependency. Here I've rewritten the protokernel in native NEON, eliminating the need for sse2neon.

Now that the `volk_8u_x4_conv_k7_r2_8u` kernel is adequately tested, we can be confident that the updated protokernel is correct.

Performance is more or less unchanged:

Old:
```
neonspiral completed in 4557.97 ms
generic completed in 29817.1 ms
```

New:
```
neonspiral completed in 4591.51 ms
generic completed in 29832.3 ms
```