@jdemel 2024-02-09
Merging this PR and waiting for #756 to be rebased. I appreciate that the code seems to simplify.